# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Codex API",
    "description": "REST API for Codex digital library server",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tasks": {
      "get": {
        "tags": [
          "Task Queue"
        ],
        "summary": "List tasks with optional filtering",
        "description": "# Permission Required\n- `tasks:read`",
        "operationId": "list_tasks",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status (pending, processing, completed, failed)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "task_type",
            "in": "query",
            "description": "Filter by task type",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit number of results (default: 50)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tasks retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Permission denied"
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "Task Queue"
        ],
        "summary": "Create a new task",
        "description": "# Permission Required\n- `tasks:write`",
        "operationId": "create_task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Permission denied"
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/tasks/nuke": {
      "delete": {
        "tags": [
          "Task Queue"
        ],
        "summary": "Nuclear option: Delete ALL tasks",
        "description": "# Permission Required\n- `admin`",
        "operationId": "nuke_all_tasks",
        "responses": {
          "200": {
            "description": "All tasks deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeTasksResponse"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied (admin only)"
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/tasks/purge": {
      "delete": {
        "tags": [
          "Task Queue"
        ],
        "summary": "Purge old completed/failed tasks",
        "description": "# Permission Required\n- `tasks:write`",
        "operationId": "purge_old_tasks",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "description": "Delete tasks older than N days (default: 30)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tasks purged successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeTasksResponse"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied"
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/tasks/stats": {
      "get": {
        "tags": [
          "Task Queue"
        ],
        "summary": "Get queue statistics",
        "description": "# Permission Required\n- `tasks:read`",
        "operationId": "get_task_stats",
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStats"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied"
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/api/v1/tasks/{id}": {
      "get": {
        "tags": [
          "Task Queue"
        ],
        "summary": "Get task by ID",
        "description": "# Permission Required\n- `tasks:read`",
        "operationId": "get_task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied"
          },
          "404": {
            "description": "Task not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/tasks/{id}/cancel": {
      "post": {
        "tags": [
          "Task Queue"
        ],
        "summary": "Cancel a task",
        "description": "# Permission Required\n- `tasks:write`",
        "operationId": "cancel_task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task cancelled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Task cannot be cancelled"
          },
          "403": {
            "description": "Permission denied"
          },
          "404": {
            "description": "Task not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/tasks/{id}/retry": {
      "post": {
        "tags": [
          "Task Queue"
        ],
        "summary": "Retry a failed task",
        "description": "# Permission Required\n- `tasks:write`",
        "operationId": "retry_task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task queued for retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Task is not in failed state"
          },
          "403": {
            "description": "Permission denied"
          },
          "404": {
            "description": "Task not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/tasks/{id}/unlock": {
      "post": {
        "tags": [
          "Task Queue"
        ],
        "summary": "Unlock a stuck task",
        "description": "# Permission Required\n- `tasks:write`",
        "operationId": "unlock_task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task unlocked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied"
          },
          "404": {
            "description": "Task not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTaskRequest": {
        "type": "object",
        "required": [
          "task_type"
        ],
        "properties": {
          "priority": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Priority level (higher = more urgent)",
            "example": 0
          },
          "scheduled_for": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When to run the task (defaults to now)",
            "example": "2024-01-15T12:00:00Z"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType",
            "description": "Type of task to create"
          }
        }
      },
      "CreateTaskResponse": {
        "type": "object",
        "required": [
          "task_id"
        ],
        "properties": {
          "task_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the created task",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        }
      },
...
      "MessageResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Task 550e8400-e29b-41d4-a716-446655440000 cancelled"
          }
        }
      },
...
      "PurgeTasksResponse": {
        "type": "object",
        "required": [
          "deleted"
        ],
        "properties": {
          "deleted": {
            "type": "integer",
            "format": "int64",
            "description": "Number of tasks deleted",
            "example": 42,
            "minimum": 0
          }
        }
      },
...
      "TaskResponse": {
        "type": "object",
        "required": [
          "id",
          "task_type",
          "status",
          "priority",
          "attempts",
          "max_attempts",
          "scheduled_for",
          "created_at"
        ],
        "properties": {
          "attempts": {
            "type": "integer",
            "format": "int32",
            "description": "Number of execution attempts",
            "example": 0
          },
          "book_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Associated book ID (if applicable)"
          },
          "completed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When task execution completed"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the task was created",
            "example": "2024-01-15T10:00:00Z"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique task identifier",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "last_error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Error message from last failed attempt"
          },
          "library_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Associated library ID (if applicable)",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "locked_by": {
            "type": [
              "string",
              "null"
            ],
            "description": "Worker ID that has locked this task",
            "example": "worker-1"
          },
          "locked_until": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When the lock expires"
          },
          "max_attempts": {
            "type": "integer",
            "format": "int32",
            "description": "Maximum number of allowed attempts",
            "example": 3
          },
          "params": {
            "description": "Task-specific parameters"
          },
          "priority": {
            "type": "integer",
            "format": "int32",
            "description": "Priority level (higher = more urgent)",
            "example": 0
          },
          "result": {
            "description": "Task execution result"
          },
          "scheduled_for": {
            "type": "string",
            "format": "date-time",
            "description": "When the task is scheduled to run",
            "example": "2024-01-15T12:00:00Z"
          },
          "series_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Associated series ID (if applicable)"
          },
          "started_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When task execution started"
          },
          "status": {
            "type": "string",
            "description": "Current status (pending, processing, completed, failed)",
            "example": "pending"
          },
          "task_type": {
            "type": "string",
            "description": "Type of task (scan_library, generate_thumbnail, etc.)",
            "example": "scan_library"
          }
        }
      },
      "TaskStats": {
        "type": "object",
        "description": "Task queue statistics",
        "required": [
          "pending",
          "processing",
          "completed",
          "failed",
          "stale",
          "total",
          "by_type"
        ],
        "properties": {
          "by_type": {
            "type": "object",
            "description": "Breakdown by task type and status",
            "additionalProperties": {
              "$ref": "#/components/schemas/TaskTypeStats"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "completed": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "failed": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "pending": {
            "type": "integer",
            "format": "int64",
            "description": "Total counts across all task types",
            "minimum": 0
          },
          "processing": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "stale": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "TaskTypeStats": {
        "type": "object",
        "description": "Statistics for a specific task type",
        "required": [
          "pending",
          "processing",
          "completed",
          "failed",
          "stale",
          "total"
        ],
        "properties": {
          "completed": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "failed": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "pending": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "processing": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "stale": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "jwt_bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/v1/auth/login"
      }
    }
  },
...
}